refactor: EAP
